Fixed Way refs (deltaCoded) and many extensions

+++ src/osmCore/COSMNode.h
@@ -10,17 +10,17 @@ class COSMNode : public CGeoPoint
 {
   using tPropertyMap = std::map<std::string, std::string>;
 public:
-  COSMNode( const int osmNodeId, const double lat, const double lon );
+  COSMNode( const int64_t osmNodeId, const double lat, const double lon );
   virtual ~COSMNode();
 
-  inline int getId() const { return m_nodeId; };
+  inline int64_t getId() const { return m_nodeId; };
 
   void addOSMNodeProperty( const std::string& name, const std::string& value);
 
   void Print() const;
 
 private:
-  int m_nodeId;
+  int64_t m_nodeId;
 
   tPropertyMap m_nodeProperties;
 };
